requirements.txt
Django==2.0.1
psycopg2==2.7.3.2
pytest==3.4.0
djangorestframework==3.7.7
